# packages/api-contracts/openapi.v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenMake.Ai API",
    "description": "\nAI 채팅 어시스턴트 API 문서\n\n## 기능\n- 채팅 메시지 전송 및 스트리밍 응답\n- 파일 업로드 및 문서 분석\n- 클러스터 관리\n- 사용자 인증\n\n## 인증\n대부분의 API는 JWT 토큰 인증이 필요합니다.\n`Authorization: Bearer <token>` 헤더를 사용하세요.\n\n## CSRF (비브라우저 클라이언트)\nBearer 헤더 요청은 CSRF 검증에서 면제됩니다. 단 **사전 인증 mutating 요청**\n(login/register/refresh)은 `GET /api/csrf-token` 으로 토큰을 발급받아\n발급 쿠키와 `X-CSRF-Token` 헤더를 함께 보내야 합니다 (Double-Submit).\n",
    "version": "1.0.0"
  },
...
  "paths": {
    "/api/chat": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "채팅 메시지 전송",
        "description": "AI에게 메시지를 전송하고 응답을 받습니다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "message"
                ],
                "properties": {
                  "message": {
                    "type": "string",
                    "description": "사용자 메시지"
                  },
                  "model": {
                    "type": "string",
                    "description": "사용할 모델",
                    "default": "default"
                  },
                  "memory": {
                    "type": "array",
                    "description": "대화 히스토리",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role": {
                          "type": "string"
                        },
                        "content": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "webSearch": {
                    "type": "boolean",
                    "description": "웹 검색 활성화",
                    "default": false
                  },
                  "thinking": {
                    "type": "boolean",
                    "description": "Sequential Thinking 활성화",
                    "default": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": {
                      "type": "string"
                    },
                    "model": {
                      "type": "string"
                    },
                    "tokensUsed": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증 필요"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
...
    "/api/v1/api-keys": {
      "post": {
        "tags": [
          "API Keys"
        ],
        "summary": "새 API Key 생성",
        "description": "새 API Key를 발급합니다. 평문 키는 이 응답에서만 반환됩니다.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 100,
                    "description": "API Key 이름"
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "scopes": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "default": [
                      "*"
                    ]
                  },
                  "allowed_models": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "default": [
                      "*"
                    ]
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API Key 생성 성공 (평문 키 포함)"
          },
          "401": {
            "description": "인증 필요"
          },
          "429": {
            "description": "키 한도 초과"
          }
        }
      },
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "API Key 목록 조회",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "API Key 목록"
          },
          "401": {
            "description": "인증 필요"
          }
        }
      }
    },
...
    "/api/v1/usage": {
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "API Key 사용량 요약",
        "description": "현재 API Key의 전체 사용량 통계를 반환합니다.",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "사용량 통계"
          },
          "401": {
            "description": "API Key 인증 필요"
          }
        }
      }
    },
...
    "/api/chat/folders": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "대화 폴더 목록(157)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "folders"
                      ],
                      "properties": {
                        "folders": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ConversationFolder"
                          }
                        }
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiMeta"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "대화 폴더 생성(157)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 64
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성됨",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "folder"
                      ],
                      "properties": {
                        "folder": {
                          "$ref": "#/components/schemas/ConversationFolder"
                        }
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiMeta"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 입력",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "409": {
            "description": "상한 초과 또는 같은 이름",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chat/sessions": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "세션 목록 조회",
        "description": "사용자의 채팅 세션 목록을 조회합니다. `?q=` 지정 시 제목+본문 검색 (결과에 snippet 포함). cursor 페이지네이션 없음 (limit-only).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            },
            "description": "최대 조회 개수"
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "제목+메시지 본문 검색어"
          },
          {
            "name": "folderId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "폴더 필터(157) — `none` 이면 미분류"
          },
          {
            "name": "tag",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "태그 필터(157)"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "sessions"
                      ],
                      "properties": {
                        "sessions": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/SessionSummary"
                          }
                        }
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiMeta"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "새 세션 생성",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "세션 제목 (선택)"
                  },
                  "model": {
                    "type": "string",
                    "description": "모델 id (선택)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "session"
                      ],
                      "properties": {
                        "session": {
                          "type": "object",
                          "required": [
                            "id",
                            "title",
                            "createdAt",
                            "updatedAt"
                          ],
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "userId": {
                              "type": "string",
                              "nullable": true
                            },
                            "anonSessionId": {
                              "type": "string",
                              "nullable": true
                            },
                            "title": {
                              "type": "string"
                            },
                            "createdAt": {
                              "type": "string"
                            },
                            "updatedAt": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiMeta"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/me/model-assignments": {
      "get": {
        "tags": [
          "Model Assignments"
        ],
        "summary": "내 모델 배정 목록",
        "description": "슬롯 목록 + 본인 배정 + 실효(effective) 해석을 반환합니다.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ModelAssignmentsResponse"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiMeta"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/me/agents": {
      "get": {
        "tags": [
          "User Agents"
        ],
        "summary": "커스텀 에이전트 목록",
        "description": "본인 소유 + 워크스페이스 공유 에이전트 목록을 반환합니다 (타인 user_id 비노출). 채팅에서 `userAgentId` 로 지정하면 산업 에이전트 자동 라우팅을 우회합니다.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "agents"
                      ],
                      "properties": {
                        "agents": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/UserAgent"
                          }
                        }
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiMeta"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiMeta": {
        "type": "object",
        "required": [
          "timestamp"
        ],
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "requestId": {
            "type": "string"
          }
        }
      },
...
      "ConversationFolder": {
        "type": "object",
        "required": [
          "id",
          "name",
          "position",
          "sessionCount",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "position": {
            "type": "integer"
          },
          "sessionCount": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key (omk_live_...) — X-API-Key 헤더 또는 Authorization: Bearer"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ]
}